marketdb/requirements.txt
# native dependencies

Pillow==10.1.0
alembic==1.8.0
beautifulsoup4==4.11.1
chardet==5.2.0
cryptography==41.0.1
db-dtypes==1.3.0
firebase-admin==5.2.0
gcp-storage-emulator==2022.6.11
gevent==23.9.1
google-cloud-bigquery==3.19.0
google-cloud-pubsub==2.18.4
google-cloud-storage==2.13.0
google.cloud.logging==3.8.0
grpcio==1.66.1
gunicorn==21.2.0
inapppy==2.5.2
inquirer==2.9.2
lxml==5.3.0
numpy==1.26.2
pandas==2.1.3
pip-tools>=7.3.0
protobuf==4.25.5
psycopg2-binary==2.9.9
pytest==7.4.3
python-dateutil==2.8.2
pyyaml==6.0.1
scipy==1.14.1
signalr-client==0.0.7
swifter==1.3.4
uvicorn==0.24.0

# Core dependencies at compatible versions
Werkzeug==2.2.2
Flask==2.0.3
Flask-CORS==3.0.10
Django==4.1.1
celery==5.2.7

# Pin these specific versions for better compatibility
pip-tools==7.3.0
setuptools>=65.5.1  # Add this if not present
wheel>=0.40.0      # Add this if not present
google-crc32c==1.3.0

# Rest of your dependencies
asgiref==3.5.2
blinker==1.5
click==8.1.3
datasketch==1.5.7
django-cors-headers==3.13.0
django-filter==22.1
django-prometheus==2.2.0
django-storages==1.12.3
djangorestframework-simplejwt==5.2.0
djangorestframework==3.13.1
drf-nested-routers==0.93.4
extruct==0.14.0
fake-headers==1.0.2
faker==24.4.0
fakeredis==2.9.2
fastapi-limiter==0.1.5
fastapi-pagination==0.9.3
fastapi==0.78.0
freezegun==1.2.2
fsspec==2024.6.1
fuzzywuzzy==0.18.0
gcsfs==2024.6.1
joblib==1.0.1
loguru==0.6.0
prometheus-client==0.16.0
pyjwt==2.4.0
python-dateutil==2.8.2
python-multipart==0.0.6
python-slugify==6.1.2
pytz==2024.1
redis==4.3.4
responses==0.22.0
schedule==1.2.0
sentry-sdk==1.9.10
six==1.16.0
sqlparse==0.4.4
stream-python==5.1.1
tqdm==4.64.0
trafilatura==1.4.0
unidecode==1.3.4
waitress==2.1.2
wget==3.2

# for dev
django-extensions==3.2.3
django-debug-toolbar==4.2.0
factory-boy==3.3.1
coverage==7.3.2
python-dotenv==1.0.0
mock==5.0.1
pylint==2.17.0
drf-yasg==1.21.10
